# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "eait",
    "version": "v1",
    "license": {
      "name": "AGPL-3.0-only",
      "identifier": "AGPL-3.0-only"
    }
  },
  "servers": [
    {
      "url": "https://api.eait.fit"
    }
  ],
  "paths": {
...
    "/v1/profile": {
      "get": {
        "operationId": "get_profile",
        "summary": "The caller's profile, targets and limits.",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Not onboarded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "patch": {
        "operationId": "patch_profile",
        "summary": "Edit the profile. Targets are recomputed; a rejected field comes back as 422.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchProfileRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Not onboarded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Rejected",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileRejected"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/v1/messages": {
      "get": {
        "operationId": "get_messages",
        "summary": "The thread, newest page first.",
        "parameters": [
          {
            "name": "before",
            "in": "query",
            "required": false,
            "description": "Page before this sequence number.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "description": "Page size.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatHistoryResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "operationId": "post_messages",
        "summary": "One turn. Charged when it reaches the model. Streams with `accept: application/x-ndjson`.",
        "parameters": [
          {
            "name": "accept",
            "in": "header",
            "required": false,
            "description": "`application/x-ndjson` streams one JSON object per line; the last line is the response body.",
            "schema": {
              "type": "string",
              "enum": [
                "application/x-ndjson",
                "application/json"
              ]
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MessageRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              },
              "application/x-ndjson": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "402": {
            "description": "Payment required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Not onboarded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Target gone",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "415": {
            "description": "Unsupported image",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Rejected",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "Upstream failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/v1/diary/day": {
      "get": {
        "operationId": "get_day",
        "summary": "One day of the diary.",
        "parameters": [
          {
            "name": "date",
            "in": "query",
            "required": false,
            "description": "YYYY-MM-DD; today by default.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DayResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Not onboarded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ChatHistoryResponse": {
        "type": "object",
        "properties": {
          "entries": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ChatEntry"
            },
            "description": "Oldest first within the page."
          },
          "before": {
            "type": [
              "number",
              "null"
            ],
            "description": "Pass back as `before` for the next older page; null once the start of the thread is in hand."
          }
        },
        "required": [
          "entries",
          "before"
        ],
        "additionalProperties": false
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        },
        "required": [
          "error"
        ],
        "additionalProperties": {},
        "description": "Every error body the API can produce, other than the refusals above."
      },
...
      "HandleTextResult": {
        "anyOf": [
          {
            "$ref": "#/components/schemas/Answered"
          },
          {
            "$ref": "#/components/schemas/MealProposed"
          },
          {
            "$ref": "#/components/schemas/MealUpdated"
          },
          {
            "$ref": "#/components/schemas/MealRedated"
          },
          {
            "$ref": "#/components/schemas/TargetGone"
          },
          {
            "$ref": "#/components/schemas/Refusal"
          }
        ]
      },
...
      "MessageResponse": {
        "$ref": "#/components/schemas/HandleTextResult"
      },
...
      "ProfileRejected": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "const": "invalid-profile"
          },
          "field": {
            "type": "string",
            "enum": [
              "lang",
              "goal",
              "sex",
              "birth_year",
              "age",
              "height_cm",
              "weight_kg",
              "target_weight_kg",
              "activity",
              "pace",
              "units",
              "struggles",
              "country",
              "restrictions",
              "diet",
              "medical",
              "medical_limitations",
              "food_allergies",
              "product_limitations",
              "complete_onboarding"
            ]
          },
          "reason": {
            "type": "string",
            "enum": [
              "target-weight-below-healthy-bmi",
              "age-below-minimum",
              "out-of-range"
            ]
          },
          "minHealthyKg": {
            "type": "number",
            "description": "Present for `target-weight-below-healthy-bmi` — what the app shows as the lowest it accepts."
          }
        },
        "required": [
          "error",
          "field",
          "reason"
        ],
        "additionalProperties": false,
        "description": "A rejected profile patch. 422, and the app re-asks the question rather than clicking past it. `target-weight-below-healthy-bmi` is the anorexia guard in `targets.ts`; it is a refusal, not a warning, because a warning is a thing the users it exists for will click past."
      },
      "ProfileResponse": {
        "type": "object",
        "properties": {
          "profile": {
            "$ref": "#/components/schemas/Profile"
          },
          "targets": {
            "$ref": "#/components/schemas/FoodTargets"
          },
          "basis": {
            "$ref": "#/components/schemas/TargetBasis",
            "description": "Why the targets are what they are. `basis.floorApplied` MUST be surfaced to the user."
          },
          "age": {
            "type": [
              "number",
              "null"
            ],
            "description": "The account's age in whole years, computed HERE from `birth_year` in this server's timezone — `ageFrom`, the same arithmetic the target model consumes.\n\nSENT, NEVER COMPUTED BY THE CLIENT: `birth_year` is a year, not a birthday, so \"current year minus it\" is already an approximation — having the server answer it keeps every surface on the same approximation instead of each client subtracting in its own zone and disagreeing by one around the new year (#97 review). `null` when the profile holds no usable birth year; a client prints nothing then, not a guess."
          },
          "onboarded": {
            "type": "boolean",
            "description": "Null until onboarding completes."
          },
          "isAdmin": {
            "type": "boolean",
            "description": "Whether this account holds the admin role (#391b).\n\nA BOOLEAN, and deliberately not the role itself: what a client needs is \"may I offer the admin section\", and a role string invites a client to reason about roles it was never told the rules for. It is also why `role` is not on `Profile` — a field there would be writable by `PATCH` in one store implementation and refused in the other.\n\nADVISORY, NEVER AN AUTHORITY. This decides whether a link is drawn. Every path under `/admin` checks the role again, server-side, on every request; a client that set this to true itself would gain a menu entry and nothing behind it."
          },
          "limits": {
            "$ref": "#/components/schemas/Limits",
            "description": "What this server will actually accept. Carried here because the profile is fetched at boot and on every refresh, so the app learns the limits of the environment it is talking to instead of assuming the ones it was compiled with."
          },
          "timezone": {
            "type": "string",
            "description": "The IANA zone this server computes calendar dates in.\n\nSent for the same reason `limits` is: it is server configuration the client must agree with rather than assume. The app aggregates health samples into days BEFORE sending them, and if it used the device's zone while the server dated meals in this one, a day's food and that same day's health would describe two different twenty-four-hour windows — on one screen, invisibly, and only for people who travel."
          },
          "entitlement": {
            "$ref": "#/components/schemas/Entitlement",
            "description": "The paid tier, as the SERVER sees it.\n\nThe store told RevenueCat, RevenueCat told this server by webhook, and this is the answer. The app has its own copy from the purchases SDK and must not branch on it for anything the server enforces: the two disagree for a few seconds after every purchase, and only one of them is the one refusing requests. `limits.dailyPhotoCap` above is already computed from this."
          },
          "pairAddress": {
            "type": "string",
            "description": "Where to send a browser to pair with this account — host and path, no scheme (#408).\n\nSENT, NEVER COMPOSED BY THE CLIENT, for the reason  {@link  API_VERSION }  states: shipped apps outlive the server they were built against. Every build before this field derived the address from its own compiled-in `API_URL` and printed `api.eait.fit/start`; that name becomes a 301 to the browser's own host and eventually stops serving the page, and a binary already on a phone cannot be told. Reading it off the profile means the next move of this surface needs no new build — which is the whole point, and is why the path is here too and not just the host.\n\nNO SCHEME, because it is read off a phone screen and typed into an address bar by hand.\n\nEMPTY when the server has been told neither origin, which is development. The client then keeps its own answer: the host it is already talking to is the right one there, and a guess invented server-side would be wrong on every worktree at once."
          },
          "telegramBot": {
            "type": [
              "string",
              "null"
            ],
            "description": "The username of this server's Telegram bot, for a Connect Telegram link, or null.\n\nNULL WHILE THE CONNECTOR IS OFF — no token, `--demo`, a dead token, or Telegram not reached yet — and a client draws no link then. The link is `https://t.me/<this>?start=<code>`, with a code from  {@link  ROUTES.authPair }  minted at the moment of the tap: a code shown in advance would die on the screen in five minutes."
          },
          "healthConnected": {
            "type": "boolean",
            "description": "Whether health sync is actually arriving: any `HealthDay` row stored in the last seven days, computed here so the You surface's \"connected\" is a fact and not a flag a client can set (#97). The phone backfills weight history on connect, so a fresh sync counts from day one."
          },
          "paywall": {
            "$ref": "#/components/schemas/WebPaywall",
            "description": "The web paywall, computed from this server's `EAIT__BACKEND__WEB_*` block (#77).\n\nSENT, NEVER COMPILED, for the same reason `limits` is: the web app can be self-hosted, so the plans, prices and exit offer are the operator's variables, and this is how the bundle learns them — already formatted in the account's language and with `{userId}` already filled into every checkout link from this account's own id, because the RevenueCat webhook grants the purchase to `app_user_id` and nothing else.\n\nPlans and `exitOffer` are null where the operator configured no checkout. ALL NULL means the host sells nothing: no paywall surface at all, whatever `entitlement` and `limits` say about this account. A server that predates this field sends no `paywall` — a client reads that as absent too, exactly as it does `entitlement`."
          },
          "coachName": {
            "type": "string",
            "description": "The coach's name in the account's language — `THREAD_COPY`'s `coach.name` — sent here because the browser bundle cannot import the Lingui table that copy lives in (#92 review): Lingui words reach a client only server-sent. Surfaces fill their `{coach}` placeholders with it. A server that predates the field sends none, and a client falls back to the copy's own name."
          },
          "hasLoggedMeal": {
            "type": "boolean",
            "description": "Whether this account has ever logged a meal — any date, not just inside the diary window (#92 review). The first-meal surfaces (Home's free-meal flow, the log's first verdict) read it as the ONE \"nothing logged yet\" answer, so two clients can never disagree about which meal was first, and neither repeats the `/v1/diary/week` probe it replaced. The other conditions of that gate — `onboarded`, `entitlement`, `limits.sampleUsed` — are already on this response."
          },
          "donate": {
            "$ref": "#/components/schemas/DonateLinks",
            "description": "{@link  DonateLinks }  — the operator's donation URLs, every one null when none are set."
          }
        },
        "required": [
          "profile",
          "targets",
          "basis",
          "age",
          "onboarded",
          "isAdmin",
          "limits",
          "timezone",
          "entitlement",
          "pairAddress",
          "telegramBot",
          "healthConnected",
          "paywall",
          "coachName",
          "hasLoggedMeal",
          "donate"
        ],
        "additionalProperties": false,
        "description": "Profile plus everything derived from it, so the app never recomputes targets locally."
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}